==> code_archive/5a_stage_offsets.Rmd <==
---
title: "ET stage offset"
author: "Megan Sears"
date: "`r Sys.Date()`"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}

knitr::opts_chunk$set(
  echo = F,
  message = F, 
  warning  = F)

# load packages
library(tidyverse)
library(lubridate)
library(here)

```

Read in stage data

```{r}

# Raw stage data - already converted to stage height
stage_raw <- read_csv(here('./data/stage/et_stage_sensors.csv')) %>%
  mutate(datetime = datetime - 25200) # only apply this if it's NOT in MST

# forze TZ
tz(stage_raw$datetime) <- 'MST'

# read in manual discharge measurements
manual_q <- read_csv(here('./data/discharge_et.csv')) %>%
  mutate(datetime = mdy_hms(paste(date, time), tz = 'MST')) %>%
  select(site, datetime, manual_stage_mean, bed_elev_adjustment_cm, discharge_Ls)

# read in field notes (manual stage, bed elevation)
field_notes <- read_csv(here('./data/field_notes/et_field_notes.csv')) %>%
  mutate(datetime = mdy_hms(paste(date, time), tz = 'MST')) %>%
  select(-c(date, time))

```

Plot all raw stage data

```{r}

hydrograph <- ggplot(stage_raw)+
  geom_line(aes(x=datetime,y=Stage_cm,color=site))

ggplotly(hydrograph)

```

# HM

```{r}

hm_adjust <- stage_raw %>%
  filter(site == 'hm') %>%
  select(datetime, Stage_cm, site)

hm_field <- field_notes %>%
  filter(site %in% c('HM', 'Hm')) %>%
  mutate(site = tolower(site))

hm_adjust <- hm_adjust %>%
  filter(!datetime < ymd_hms('2022-06-20 10:55:00', tz = 'MST'),
         datetime < ymd_hms('2022-07-12 13:05:00', tz = 'MST') | 
         datetime > ymd_hms('2022-07-12 15:35:00', tz = 'MST'),
         datetime < ymd_hms('2022-07-21 13:00:00', tz = 'MST') | 
         datetime > ymd_hms('2022-07-21 13:35:00', tz = 'MST'),
         !datetime == ymd_hms('2022-08-17 12:35:00', tz = 'MST'),
         !datetime == ymd_hms('2023-09-28 11:50:00', tz = 'MST'),
         datetime < ymd_hms('2023-05-24 11:00:00', tz = 'MST') | 
         datetime > ymd_hms('2023-05-24 14:45:00', tz = 'MST'),
         !datetime == ymd_hms('2023-09-28 11:55:00', tz = 'MST'),
         !datetime == ymd_hms('2023-07-27 08:00:00', tz = 'MST')) %>%
  mutate(Stage_fix = Stage_cm - 17.5,
         Stage_fix = if_else(datetime >= ymd_hms('2022-07-05 19:00:00', tz = 'MST'),
                             Stage_fix + 6,
                             Stage_fix),
         Stage_fix = if_else(datetime >= ymd_hms('2022-08-04 12:35:00', tz = 'MST'),
                             Stage_fix - 1.5,
                             Stage_fix))

hm <- ggplot(hm_adjust)+
  geom_line(aes(x=datetime,y=Stage_fix)) +
  #geom_point(aes(x=datetime,y=Stage_fix)) +
  geom_line(aes(x=datetime, y= Stage_cm), color = 'grey') +
  geom_point(data = hm_field, aes(x=datetime, y=manual_stage_adjusted), color = 'red')

ggplotly(hm)

# get manual Qs
q_manual_hm <- manual_q %>%
  filter(site == 'hm')

# lets apply rating curve!!
hm_adjust <- hm_adjust %>%
  mutate(Q_Ls = (1.515791 + 0.2351925*Stage_fix)^2)

hm_q <- ggplot(hm_adjust)+
  geom_line(aes(x=datetime,y=Q_Ls)) +
  geom_point(data = q_manual_hm, aes(x=datetime, y=discharge_Ls), color = 'blue')

ggplotly(hm_q)

q_manual_hm <- left_join(q_manual_hm, hm_adjust, by = 'datetime')

hm_r2 <- cor(q_manual_hm$discharge_Ls, q_manual_hm$Q_Ls, use = 'complete.obs')^2
#0.88

```

# HUM - BAD

cannot use rating curve here

```{r}

hum_adjust <- stage_raw %>%
  filter(site == 'hum') %>%
  select(datetime, Stage_cm, site)

hum_field <- field_notes %>%
  filter(site %in% c('HUM', 'HUm')) %>%
  mutate(site = tolower(site))

hum_adjust <- hum_adjust %>%
  filter(datetime < ymd_hms('2022-07-12 11:50:00', tz = 'MST') | 
         datetime > ymd_hms('2022-07-12 15:35:00', tz = 'MST'),
         !datetime < ymd_hms('2022-06-20 10:20:00', tz = 'MST'),
         !datetime == ymd_hms('2022-07-21 13:55:00', tz = 'MST'),
         !datetime == ymd_hms('2022-08-04 12:20:00', tz = 'MST'),
         !datetime == ymd_hms('2022-10-18 13:25:00', tz = 'MST'),
         !datetime == ymd_hms('2023-05-24 10:55:00', tz = 'MST'),
         !datetime == ymd_hms('2023-07-19 10:15:00', tz = 'MST')) %>% 
  filter(datetime < ymd_hms('2023-05-24 10:00:00', tz = 'MST') | 
         datetime > ymd_hms('2023-05-24 15:25:00', tz = 'MST')) %>%
  mutate(Stage_fix = Stage_cm - 4.3,
         Stage_fix = if_else(datetime >= ymd_hms('2022-07-14 23:10:00', tz = 'MST'),
                             Stage_fix - 4.5,
                             Stage_fix),
         Stage_fix = if_else(datetime >= ymd_hms('2022-07-21 14:00:00', tz = 'MST'),
                             Stage_fix + 6,
                             Stage_fix),
         Stage_fix = if_else(datetime >= ymd_hms('2022-07-28 17:50:00', tz = 'MST'),
                             Stage_fix - 7,
                             Stage_fix),
         Stage_fix = if_else(datetime >= ymd_hms('2022-08-04 13:22:00', tz = 'MST'),
                             Stage_fix - 3,
                             Stage_fix),
         Stage_fix = if_else(datetime >= ymd_hms('2022-08-15 19:40:00', tz = 'MST'),
                             Stage_fix - 6,
                             Stage_fix),
         Stage_fix = if_else(datetime >= ymd_hms('2022-08-17 13:55:00', tz = 'MST'),
                             Stage_fix + 2,
                             Stage_fix),
         Stage_fix = if_else(datetime >= ymd_hms('2022-11-18 12:00:00', tz = 'MST'),
                             Stage_fix + 18.5,
                             Stage_fix),
         Stage_fix = if_else(datetime >= ymd_hms('2023-05-24 10:30:00', tz = 'MST'),
                             Stage_fix - 25,
                             Stage_fix),
         Stage_fix = if_else(datetime >= ymd_hms('2023-07-19 10:05:00', tz = 'MST'),
                             Stage_fix + 11.5,
                             Stage_fix)) %>%
  filter(!datetime == ymd_hms('2023-07-19 10:05:00', tz = 'MST'),
         !datetime == ymd_hms('2023-07-19 10:10:00', tz = 'MST')) %>%
  mutate(Stage_fix = if_else(datetime >= ymd_hms('2023-07-27 09:35:00', tz = 'MST'),
                             Stage_fix - 6,
                             Stage_fix),
         Stage_fix = if_else(datetime >= ymd_hms('2023-08-24 17:20:00', tz = 'MST'),
                             Stage_fix - 9,
                             Stage_fix))


hum <- ggplot(hum_adjust)+
  geom_line(aes(x=datetime,y=Stage_fix)) +
  #geom_point(aes(x=datetime,y=Stage_fix)) +
  geom_line(aes(x=datetime, y= Stage_cm), color = 'grey') +
  geom_point(data = hum_field, aes(x=datetime, y=manual_stage_adjusted), color = 'red')

ggplotly(hum)

# get manual Qs
q_manual_hum <- manual_q %>%
  filter(site == 'hum')

# lets apply rating curve!!
hum_adjust <- hum_adjust %>%
  mutate(Q_Ls = (-7.251958 + 0.3831334*Stage_fix)^2)

hum_q <- ggplot(hum_adjust)+
  geom_line(aes(x=datetime,y=Q_Ls)) +
  geom_point(data = q_manual_hum, aes(x=datetime, y=discharge_Ls), color = 'blue')

ggplotly(hum_q)

q_manual_hum <- left_join(q_manual_hum, hum_adjust, by = 'datetime')

hum_r2 <- cor(q_manual_hum$discharge_Ls, q_manual_hum$Q_Ls, use = 'complete.obs')^2
#0.31

```

# MPM - GOOD

```{r}

mpm_adjust <- stage_raw %>%
  filter(site == 'mpm') %>%
  select(datetime, Stage_cm, site)

mpm_field <- field_notes %>%
  filter(site %in% c('MPm', 'MPM')) %>%
  mutate(site = tolower(site))

mpm_adjust <- mpm_adjust %>%
  filter(!datetime < ymd_hms('2022-06-20 12:10:00', tz = 'MST'),
         !datetime == ymd_hms('2022-07-12 12:55:00', tz = 'MST'),
         !datetime == ymd_hms('2022-07-21 11:15:00', tz = 'MST'),
         !datetime == ymd_hms('2022-07-12 13:10:00', tz = 'MST')) %>%
  mutate(Stage_fix = Stage_cm + 6.4,
         Stage_fix = if_else(datetime >= ymd_hms('2022-07-05 15:15:00', tz = 'MST'),
                             Stage_fix - 4.5,
                             Stage_fix),
         Stage_fix = if_else(datetime >= ymd_hms('2022-09-05 11:20:00', tz = 'MST'),
                             Stage_fix - 1,
                             Stage_fix),
        Stage_fix = if_else(datetime >= ymd_hms('2023-03-01 12:00:00', tz = 'MST'),
                             Stage_fix + 12,
                             Stage_fix),
        Stage_fix = if_else(datetime >= ymd_hms('2023-06-29 09:40:00', tz = 'MST'),
                             Stage_fix - 4,
                             Stage_fix),
        Stage_fix = if_else(datetime >= ymd_hms('2023-07-01 16:40:00', tz = 'MST'),
                             Stage_fix + 2.5,
                             Stage_fix),
        Stage_fix = if_else(datetime >= ymd_hms('2023-07-20 05:30:00', tz = 'MST'),
                             Stage_fix + 4.5,
                             Stage_fix),
        Stage_fix = if_else(datetime >= ymd_hms('2023-08-02 17:15:00', tz = 'MST'),
                             Stage_fix - 8.5,
                             Stage_fix))
         
mpm <- ggplot(mpm_adjust)+
  geom_line(aes(x=datetime,y=Stage_fix)) +
  #geom_point(aes(x=datetime,y=Stage_fix)) +
  #geom_line(aes(x=datetime, y= Stage_cm), color = 'grey') +
  geom_point(data = mpm_field, aes(x=datetime, y=manual_stage_adjusted), color = 'red')

ggplotly(mpm)

# get manual Qs
q_manual_mpm <- manual_q %>%
  filter(site == 'mpm')

# lets apply rating curve!!
mpm_adjust <- mpm_adjust %>%
  mutate(Q_Ls = (-1.619204 + 0.6325591*Stage_fix)^2)

mpm_q <- ggplot(mpm_adjust)+
  geom_line(aes(x=datetime,y=Q_Ls)) +
  #geom_point(aes(x=datetime,y=Q_Ls)) +
  geom_point(data = q_manual_mpm, aes(x=datetime, y=discharge_Ls), color = 'blue')

ggplotly(mpm_q)

q_manual_mpm <- left_join(q_manual_mpm, mpm_adjust, by = 'datetime')

mpm_r2 <- cor(q_manual_mpm$discharge_Ls, q_manual_mpm$Q_Ls, use = 'complete.obs')^2
#0.86

```

# MUB - GOOD

go back

```{r}

mub_adjust <- stage_raw %>%
  filter(site == 'mub') %>%
  select(datetime, Stage_cm, site)

mub_field <- field_notes %>%
  filter(site %in% c('MUb', 'MUB')) %>%
  mutate(site = tolower(site))

mub_adjust2 <- mub_adjust %>%
  filter(!datetime < ymd_hms('2022-06-20 13:50:00', tz = 'MST'),
         !datetime == ymd_hms('2022-08-04 09:15:00', tz = 'MST'),
         !datetime == ymd_hms('2022-08-17 09:40:00', tz = 'MST'),
         !datetime == ymd_hms('2022-10-18 10:20:00', tz = 'MST'),
         !datetime == ymd_hms('2023-08-09 10:05:00', tz = 'MST'),
         !datetime == ymd_hms('2023-09-28 09:35:00', tz = 'MST')) %>%
  mutate(Stage_fix = Stage_cm + 4.3,
         Stage_fix = if_else(datetime >= ymd_hms('2022-06-26 22:40:00', tz = 'MST'),
                             Stage_fix - 2.5,
                             Stage_fix),
         Stage_fix = if_else(datetime >= ymd_hms('2022-08-28 20:00:00', tz = 'MST'),
                             Stage_fix + 2,
                             Stage_fix),
         Stage_fix = if_else(datetime >= ymd_hms('2023-04-01 12:00:00', tz = 'MST'),
                             Stage_fix - 2.5,
                             Stage_fix))
         

mub <- ggplot(mub_adjust2)+
  geom_line(aes(x=datetime,y=Stage_fix)) +
  #geom_point(aes(x=datetime,y=Stage_fix)) +
  #geom_line(aes(x=datetime, y= Stage_cm), color = 'grey') +
  geom_point(data = mub_field, aes(x=datetime, y=manual_stage_adjusted), color = 'blue')

ggplotly(mub)

# get manual Qs
q_manual_mub <- manual_q %>%
  filter(site == 'mub')

# lets apply rating curve!!
mub_adjust2 <- mub_adjust2 %>%
  mutate(Q_Ls = (-6.664736 + 0.7205008*Stage_fix)^2)

mub_q <- ggplot(mub_adjust2)+
  geom_line(aes(x=datetime,y=Q_Ls)) +
  geom_point(data = q_manual_mub, aes(x=datetime, y=discharge_Ls), color = 'blue')

ggplotly(mub_q)

q_manual_mub <- left_join(q_manual_mub, mub_adjust2, by = 'datetime')

mub_r2 <- cor(q_manual_mub$discharge_Ls, q_manual_mub$Q_Ls, use = 'complete.obs')^2
#0.86

```


# MUM - BAD, need to redo RC

```{r}

mum_adjust <- stage_raw %>%
  filter(site == 'mum') %>%
  select(datetime, Stage_cm, site)

mum_field <- field_notes %>%
  filter(site %in% c('MUM', 'MUm')) %>%
  mutate(site = tolower(site))

mum_adjust <- mum_adjust %>%
  filter(!datetime < ymd_hms('2022-06-20 12:30:00', tz = 'MST'),
         !datetime == ymd_hms('2022-07-12 12:15:00', tz = 'MST'),
         !datetime == ymd_hms('2022-09-05 11:00:00', tz = 'MST')) %>%
  mutate(Stage_fix = Stage_cm + 11.7,
         Stage_fix = if_else(datetime >= ymd_hms('2022-07-05 17:05:00', tz = 'MST'),
                             Stage_fix - 3,
                             Stage_fix),
         Stage_fix = if_else(datetime >= ymd_hms('2023-01-01 12:00:00', tz = 'MST'),
                             Stage_fix - 3.5,
                             Stage_fix),
         Stage_fix = if_else(datetime >= ymd_hms('2023-05-24 12:10:00', tz = 'MST'),
                             Stage_fix + 7,
                             Stage_fix),
        Stage_fix = if_else(datetime >= ymd_hms('2023-06-03 15:30:00', tz = 'MST'),
                             Stage_fix - 3,
                             Stage_fix),
        Stage_fix = if_else(datetime >= ymd_hms('2023-06-07 13:50:00', tz = 'MST'),
                             Stage_fix - 8,
                             Stage_fix),
        Stage_fix = if_else(datetime >= ymd_hms('2023-08-09 11:05:00', tz = 'MST'),
                             Stage_fix +2,
                             Stage_fix),
        Stage_fix = if_else(datetime >= ymd_hms('2023-08-25 10:40:00', tz = 'MST'),
                             Stage_fix - 5,
                             Stage_fix))






mum <- ggplot(mum_adjust) +
  geom_line(aes(x=datetime,y=Stage_fix)) +
  #geom_point(aes(x=datetime,y=Stage_fix)) +
  #geom_line(aes(x=datetime, y= Stage_cm), color = 'grey') +
  geom_point(data = mum_field, aes(x=datetime, y=manual_stage_adjusted), color = 'red')

ggplotly(mum)

# get manual Qs
q_manual_mum <- manual_q %>%
  filter(site == 'mum')

# lets apply rating curve!!
mum_adjust <- mum_adjust %>%
  mutate(Q_Ls = (-15.90045 + 0.6671579*Stage_fix)^2)

mum_q <- ggplot(mum_adjust)+
  geom_line(aes(x=datetime,y=Q_Ls)) +
  geom_point(data = q_manual_mum, aes(x=datetime, y=discharge_Ls), color = 'blue')

ggplotly(mum_q)

q_manual_mum <- left_join(q_manual_mum, mum_adjust, by = 'datetime')

mum_r2 <- cor(q_manual_mum$discharge_Ls, q_manual_mum$Q_Ls, use = 'complete.obs')^2
# NEED TO REDO RATING CURVE

```

# P1 - GREAT

```{r}

p1_adjust <- stage_raw %>%
  filter(site == 'p1') %>%
  select(datetime, Stage_cm, site)

p1_field <- field_notes %>%
  filter(site %in% c('P1', 'p1')) %>%
  mutate(site = tolower(site))

p1_adjust <- p1_adjust %>%
  filter(datetime < ymd_hms('2022-07-12 08:10:00', tz = 'MST') | 
         datetime > ymd_hms('2022-07-12 15:35:00', tz = 'MST'),
         !datetime == ymd_hms('2023-07-19 08:40:00', tz = 'MST')) %>%
  mutate(Stage_fix = Stage_cm + 5.2,
         Stage_fix = ifelse(datetime >= ymd_hms('2022-07-12 09:10:00', tz = 'MST'),
                            Stage_fix - 3,
                            Stage_fix),
         Stage_fix = ifelse(datetime >= ymd_hms('2022-07-29 12:00:00', tz = 'MST'),
                            Stage_fix + 4,
                            Stage_fix),
         Stage_fix = ifelse(datetime >= ymd_hms('2022-08-15 14:40:00', tz = 'MST'),
                            Stage_fix + 2,
                            Stage_fix),
         Stage_fix = ifelse(datetime >= ymd_hms('2022-08-26 18:00:00', tz = 'MST'),
                            Stage_fix - 2,
                            Stage_fix),
         Stage_fix = ifelse(datetime >= ymd_hms('2023-06-15 07:20:00', tz = 'MST'),
                            Stage_fix - 4,
                            Stage_fix),
         Stage_fix = ifelse(datetime >= ymd_hms('2023-07-19 08:10:00', tz = 'MST'),
                            Stage_fix + 7,
                            Stage_fix),
         Stage_fix = ifelse(datetime >= ymd_hms('2023-07-27 07:50:00', tz = 'MST'),
                            Stage_fix - 3,
                            Stage_fix),
         Stage_fix = ifelse(datetime >= ymd_hms('2023-08-25 02:00:00', tz = 'MST'),
                            Stage_fix - 3.5,
                            Stage_fix))
         


        
p1 <- ggplot(p1_adjust)+
  geom_line(aes(x=datetime,y=Stage_fix)) +
  #geom_point(aes(x=datetime,y=Stage_fix)) +
  #geom_line(aes(x=datetime, y= Stage_cm), color = 'grey') +
  geom_point(data = p1_field, aes(x=datetime, y=manual_stage_adjusted), color = 'blue')


ggplotly(p1)

# get manual Qs
q_manual_p1 <- manual_q %>%
  filter(site == 'p1')

# lets apply rating curve!!
p1_adjust <- p1_adjust %>%
  mutate(Q_Ls = (-3.12188 + 0.2616076*Stage_fix)^2)

p1_q <- ggplot(p1_adjust)+
  geom_line(aes(x=datetime,y=Q_Ls)) +
  geom_point(data = q_manual_p1, aes(x=datetime, y=discharge_Ls), color = 'blue')

ggplotly(p1_q)

q_manual_p1 <- left_join(q_manual_p1, p1_adjust, by = 'datetime')

p1_r2 <- cor(q_manual_p1$discharge_Ls, q_manual_p1$Q_Ls, use = 'complete.obs')^2
#0.96

```


# P2 - OK

```{r}

p2_adjust <- stage_raw %>%
  filter(site == 'p2') %>%
  select(datetime, Stage_cm, site)

p2_field <- field_notes %>%
  filter(site %in% c('p2', 'P2')) %>%
  mutate(site = tolower(site))

p2_adjust <- p2_adjust %>%
  filter(!datetime == ymd_hms('2022-07-21 14:25:00', tz = 'MST'),
         !datetime == ymd_hms('2022-08-04 12:50:00', tz = 'MST'),
         !datetime == ymd_hms('2023-05-24 10:05:00', tz = 'MST'),
         datetime < ymd_hms('2022-07-12 08:20:00', tz = 'MST') | 
         datetime > ymd_hms('2022-07-12 15:35:00', tz = 'MST')) %>%
  mutate(Stage_fix = ifelse(datetime >= ymd_hms('2022-07-07 15:35:00', tz = 'MST'),
                            Stage_cm - 2,
                            Stage_cm),
         Stage_fix = ifelse(datetime >= ymd_hms('2022-07-12 10:20:00', tz = 'MST'),
                            Stage_fix - 2,
                            Stage_fix),
         Stage_fix = ifelse(datetime >= ymd_hms('2022-07-14 19:35:00', tz = 'MST'),
                            Stage_fix - 7,
                            Stage_fix),
         Stage_fix = ifelse(datetime >= ymd_hms('2022-07-28 19:00:00', tz = 'MST'),
                            Stage_fix + 5,
                            Stage_fix),
         Stage_fix = ifelse(datetime >= ymd_hms('2022-08-04 13:00:00', tz = 'MST'),
                            Stage_fix - 2,
                            Stage_fix),
         Stage_fix = ifelse(datetime >= ymd_hms('2023-04-15 23:00:00', tz = 'MST'),
                            Stage_fix + 9,
                            Stage_fix),
         Stage_fix = ifelse(datetime >= ymd_hms('2023-06-07 14:00:00', tz = 'MST'),
                            Stage_fix - 4,
                            Stage_fix),
         Stage_fix = ifelse(datetime >= ymd_hms('2023-07-27 08:50:00', tz = 'MST'),
                            Stage_fix - 3,
                            Stage_fix),
         Stage_fix = ifelse(datetime >= ymd_hms('2023-08-25 06:00:00', tz = 'MST'),
                            Stage_fix - 3,
                            Stage_fix))

p2 <- ggplot(p2_adjust)+
  geom_line(aes(x=datetime,y=Stage_fix)) +
  #geom_point(aes(x=datetime,y=Stage_fix)) +
  #geom_line(aes(x=datetime, y= Stage_cm), color = 'grey') +
  geom_point(data = p2_field, aes(x=datetime, y=manual_stage_adjusted), color = 'blue')

ggplotly(p2)

# get manual Qs
q_manual_p2 <- manual_q %>%
  filter(site == 'p2')

# lets apply rating curve!!
p2_adjust <- p2_adjust %>%
  mutate(Q_Ls = if_else(datetime < ymd_hms('2023-05-24 12:00:00', tz = 'MST'), 
                                           (4.4496929 + 0.27665459*Stage_fix)^2,
                        (-6.446525 + 0.7166015*Stage_fix)^2))

p2_q <- ggplot(p2_adjust)+
  geom_line(aes(x=datetime,y=Q_Ls)) +
  geom_point(data = q_manual_p2, aes(x=datetime, y=discharge_Ls), color = 'blue')

ggplotly(p2_q)

q_manual_p2 <- left_join(q_manual_p2, p2_adjust, by = 'datetime')

p2_r2 <- cor(q_manual_p2$discharge_Ls, q_manual_p2$Q_Ls, use = 'complete.obs')^2
# 0.78

```

# combine all discharge data for sites and export to data folder

```{r}

et_estimated_discharge <- bind_rows(hm_adjust, hum_adjust,
                                    mpm_adjust, mub_adjust,
                                    mum_adjust, p1_adjust,
                                    p2_adjust)

write.csv(et_estimated_discharge, '/Users/megansears/Documents/Repos/post-fire_rainfall/data/et_estimated_q.csv')

```

# CPF prep

```{r}

# Manual Q, field notes, raw stage data
aspen_stage <- read_csv(here('./data/stage/Aspen_stage_composite.csv')) %>%
  mutate(datetime = mdy_hm(datetime), tz = 'MST') %>%
  mutate(datetime = force_tz(datetime, tzone = 'MST')) %>%
  filter(datetime > ymd_hms('2021-05-01 00:00:00', tz = 'MST'),
         datetime < ymd_hms('2023-10-01 00:00:00', tz = 'MST'))

dadd_stage <- read_csv(here('./data/stage/Dadd_stage_composite.csv')) %>%
  select(-4) %>%
  mutate(datetime = mdy_hm(datetime), tz = 'MST') %>%
  mutate(datetime = force_tz(datetime, tzone = 'MST')) %>%
  filter(datetime > ymd_hms('2021-05-01 00:00:00', tz = 'MST'),
         datetime < ymd_hms('2023-10-01 00:00:00', tz = 'MST'))

mtcamp <- read_csv(here('./data/stage/mtcampus_stage_composite.csv')) %>%
  rename(datetime = TIMESTAMP) %>% 
  mutate(datetime = mdy_hm(datetime), tz = 'MST') %>%
  mutate(datetime = force_tz(datetime, tzone = 'MST')) %>%
  filter(datetime > ymd_hms('2021-05-01 00:00:00', tz = 'MST'),
         datetime < ymd_hms('2023-10-01 00:00:00', tz = 'MST'))

# manual Q
cpf_manual_q <- read_csv(here('./data/stage/cpf_discharge_google.csv')) %>%
  filter(site %in% c('dadd', 'mtcampus', 'aspen'))


```


# Dadd

```{r}

dadd_stage_raw <- ggplot(dadd_stage)+
  geom_line(aes(x=datetime,y=Stage_mm)) +
  geom_point(aes(x=datetime,y=Stage_mm)) 

ggplotly(dadd_stage_raw)

Dadd_stage_hourly$Stage_corr_cm=ifelse(Dadd_stage_hourly$DateTime>'2021-05-18 14:30:00',
                                       Dadd_stage_hourly$Stage_corr_cm+0.6,Dadd_stage_hourly$Stage_corr_cm)

Dadd_stage_hourly$Stage_corr_cm=ifelse(Dadd_stage_hourly$DateTime>'2021-06-14 11:00:00',
                                       Dadd_stage_hourly$Stage_corr_cm+0.9,Dadd_stage_hourly$Stage_corr_cm)

```


# Aspen

# Mt. Campus
