==> code_archive/7a_response_modeling.Rmd <==
---
title: "Response modeling"
author: "Megan Sears"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: yes
    toc_float: yes
    theme: journal
editor_options:
  chunk_output_type: console
---

```{r, include = F}

knitr::opts_chunk$set(echo = F,
                      message = F,
                      fig.width = 12,
                      fig.height = 6)

library(tidyverse)
library(stringr)
library(lubridate)
library(plotly)
library(kableExtra)
library(here)
library(ggplot2); theme_set(theme_bw(base_size = 16))
library(zoo)
library(corrplot)
library(stats)
library(ggcorrplot)

```

# Read in response data

```{r}

#read in response all and drop some columns that aren't needed
response_all <- read_csv(here('data/response_all.csv')) %>%
  select(-c(1, endtime_plus12, starttime_minus1,
         crossover, response,
         treatment, burn,
         area, cat)) %>%
  mutate(response_manual = if_else(response_manual == 'Y', # change Y/N to 0/1
                                   1, 0)) %>%
  rename(stage_cm_norm_area = stage_area) %>%
  arrange(start_time) %>%
  group_by(site, year) %>%
  mutate(time_since_laststorm_hr = as.numeric(start_time - lag(start_time)), # add time since last storm (soil moist proxy)
         response_p10 = if_else(event_sum_mm > 10 & 
                                  response_manual == 1,
                                1, 0),
         response_stage4 = if_else(stage_rise_cm > 4,
                                   1,
                                   0))
```

# Bring in the GIS variables

```{r}

# read in the GIS inputs csv
gis <- read_csv(here('data/gis_inputs.csv')) %>%
  select(-Total_area_km2) %>% # drop area here because it's included already
  rename(site = ID)

# join to response_all csv and remove NDVI columns (need to put as one column)
response_all <- left_join(response_all, gis, by = 'site') %>%
  select(-c(NDVI_2021, NDVI_2022, NDVI_2023))

# make NDVI data long
ndvi <- gis %>%
  select(site, NDVI_2021, NDVI_2022, NDVI_2023) %>%
  pivot_longer(!site, names_to = 'year', values_to = 'NDVI') %>%
  mutate(year = as.numeric(sub("^NDVI_", "", year)))

# join NDVI with response_all df
response_all <- left_join(response_all, ndvi, by = c('site', 'year')) %>%
  ungroup() %>%
  mutate(fire = ifelse(fire == 'etf',
                        2,
                        1),
         geology = ifelse(geology == 'ig_met',
                           1,
                           2)) %>%  # made fire (etf [2] and cpf [1]) and geology numeric
  unite(event_IDs, site, event, year, sep = "_",
        remove = F) %>%
  mutate(site_numeric = as.numeric(as.factor(site)))
  

write.csv(response_all, 'model_variables.csv')

findsite <- response_all %>%
  distinct(site, site_numeric)


```

# Look at some of the data

```{r}
# select values that are numeric / values for histograms & boxplots
select_numeric <- response_all %>%
  select(ID, event, event_sum_mm, MI10_mmhr,
         MI20_mmhr, MI30_mmhr, MI60_mmhr,
         duration_hr, stage_rise_cm, lag2p_hr,
         response_manual, stage_cm_norm_area,
         time_since_laststorm_hr, response_p10)

select_numeric <- select_numeric %>% 
  pivot_longer(!ID, names_to = 'predictor', values_to = 'value')

# function to plot histograms of above variables
plot_hist <- function(pred) {
  select_numeric %>%
    filter(predictor == pred) %>%
    ggplot(aes(x = value)) + #,
               #shape = ID, 
               #color = ID)) + 
    geom_histogram() + 
    ggtitle(pred) +
    theme_bw()
}

name <- unique(select_numeric$predictor)

lapply(name, plot_hist)

###################################

# boxplot by site for above variables
plot_box <- function(pred) {
  select_numeric %>%
    filter(predictor == pred) %>%
    ggplot(aes(x = ID, 
               y = value)) + 
    geom_boxplot() + 
    ggtitle(pred) +
    theme_bw()
}

name <- unique(select_numeric$predictor)

lapply(name, plot_box) # ignore event and response (0/1) boxplots

# MI60 vs stage rise
stage <- ggplot(response_all, aes(MI60_mmhr, stage_rise_cm)) + geom_point()

ggplotly(stage)

```

# Correlations for predictors 

```{r}

# correlation of predictors
cor_pred <- response_all %>%
  select(-c(event,
            ID,
            start_time,
            end_time,
            datetime_maxstage,
            Stage_cm_max,
            stage_rise_cm,
            lag2p_hr,
            response_manual,
            stage_cm_norm_area,
            response_p10,
            response_stage4,
            site))

matrix_cor_pred <- cor(cor_pred, use = 'complete.obs')

corrplot(matrix_cor_pred, method = "color") # hard to look at

ggcorrplot(matrix_cor_pred,
           lab = T) # add cor # on plot

ggcorrplot(matrix_cor_pred,
           lab = F) # no cor # on plot

# look at predictors with response variables
cor_both <- response_all %>%
  select(-c(event,
            ID,
            start_time,
            end_time,
            datetime_maxstage,
            Stage_cm_max))

matrix_cor_all <- cor(cor_both, use = 'complete.obs')

ggcorrplot(matrix_cor_all,
           lab = T)

# duration hr seems to be most correlated with stage response variables 
# variables to keep based on questions: dnbr, ndvi, %treated, area, topo variable

```
