==> code_archive/7b_RF_modeling.Rmd <==
---
title: "Create figures"
author: "Megan Sears"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: yes
    toc_float: yes
    theme: journal
editor_options:
  chunk_output_type: console
---

# Packages
```{r}

library(pacman)
p_load(tidyverse, reshape2, here, lme4, randomForest, boot, MuMIn, magrittr)
library(ggplot2); theme_set(theme_bw(base_size = 16))

```

# Read in data

```{r}

allData <- read_csv(here('model_variables_update.csv'))

custom_labels <- c(MI60_mmhr = 'MI60 (mm/hr)', 
                   last_storm_hr = 'Hours since last storm',
                   dNBR_mean = 'Mean dNBR',
                   percent_treated = '% area mulching',
                   clay_mean = 'Mean percent clay',
                   year = 'Year',
                   ID = 'Site',
                   sand_mean = 'Mean percent sand',
                   Area_km2 = expression('Area (km'^'2'~')'),
                   Mean_slope = "Mean slope",
                   Ksat_mean = 'Mean Ksat',
                   snow_persi = 'Snow persistence')



```

# Stage rise ~ all sites

```{r}

#subset data only to dependent variable and predictors
srData <- allData %>%
  rename(percent_treated = `%treated`) %>%
  subset(., response_manual == 1, # subset data to observed stage rise events
         select = c(stage_rise_cm, ID, year, MI60_mmhr, Area_km2, last_storm_hr, 
                    dNBR_mean, sand_mean, clay_mean, percent_treated, NDVI, 
                    snow_persi, Ksat_mean, Mean_slope)) %>%
  na.omit # remove NAs, RF doesn't take any NAs


# RF model
rf_sr_all <- randomForest(stage_rise_cm ~., 
                          mtry=3, ntree=1000, # mtry started as 3
                          data=srData)


# check model performance for % variance explained
rf_sr_all # note % Var explained isn't super high (~38%), that's fine, we're not trying to produce a predictive model, just explain the data

#view crude graph of variable importance rankings
varImpPlot(rf_sr_all) 

# create more customaizable/interpretable variable importance rankings
varImps_sr <- rf_sr_all$importance %>%
  as.data.frame # save increased node purity (metric of variable importance) to data frame

stage_rise_all_RF <- varImps_sr %>%
  mutate(., normINP = IncNodePurity/max(IncNodePurity)) %>% # IncNodePurity is the metric for regression RFs (continous data)
  rownames_to_column('Predictor') %>%
  mutate(Predictor = as.factor(Predictor)) %>%
  ggplot(aes(x=reorder(Predictor, normINP), y=normINP)) + theme_bw(base_size = 20) +
      geom_bar(stat='identity', color='black', fill='black') +
      coord_flip() +
      labs(y='Normalized Variable Importance') +
      theme(axis.title.y = element_blank()) +
  scale_x_discrete(labels = custom_labels) +
      #ggtitle('Factors influencing stage rise (cm) across fires') +
  theme(legend.position="bottom",
        text = element_text(size=30,
                            family = 'Times New Roman'))

ggsave(here('figures/stage_rise_all.jpg'), stage_rise_all_RF, dpi=800)

```

# Stage rise ~ Benn

```{r}

#subset data only to dependent variable and predictors
srData_benn <- allData %>%
  rename(percent_treated = `%treated`) %>%
  filter(fire == 1) %>%
  subset(., response_manual == 1, # subset data to observed stage rise events
         select = c(stage_rise_cm, ID, year, MI60_mmhr, Area_km2, last_storm_hr, 
                    dNBR_mean, sand_mean, clay_mean, percent_treated, NDVI, 
                    snow_persi, Ksat_mean, Mean_slope)) %>%
  na.omit # remove NAs, RF doesn't take any NAs


# RF model
rf_sr_benn <- randomForest(stage_rise_cm ~., 
                          mtry=3, ntree=1000, # mtry started as 3
                          data=srData_benn)


# check model performance for % variance explained
rf_sr_benn # note % Var explained isn't super high (~38%), that's fine, we're not trying to produce a predictive model, just explain the data

#view crude graph of variable importance rankings
varImpPlot(rf_sr_benn) 

# create more customaizable/interpretable variable importance rankings
varImps_sr_benn <- rf_sr_benn$importance %>%
  as.data.frame # save increased node purity (metric of variable importance) to data frame

stage_rise_benn_RF <- varImps_sr_benn %>%
  mutate(., normINP = IncNodePurity/max(IncNodePurity)) %>% # IncNodePurity is the metric for regression RFs (continous data)
  rownames_to_column('Predictor') %>%
  mutate(Predictor = as.factor(Predictor)) %>%
  ggplot(aes(x=reorder(Predictor, normINP), y=normINP)) + theme_bw(base_size = 20) +
      geom_bar(stat='identity', color='black', fill='grey62') +
      coord_flip() +
  scale_x_discrete(labels = custom_labels) +
      labs(y='Normalized Variable Importance') +
      theme(axis.title.y = element_blank()) 
      #ggtitle('CP Factors influencing stage rise (cm) across fires')

stage_rise_benn_RF

#ggsave(here('figures/stage_rise_benn.jpg'), stage_rise_benn_RF, dpi=800)

```

# Stage rise ~ ET

```{r}

#subset data only to dependent variable and predictors
srData_et <- allData %>%
  rename(percent_treated = `%treated`) %>%
  filter(fire == 1) %>%
  subset(., response_manual == 1, # subset data to observed stage rise events
         select = c(stage_rise_cm, ID, year, MI60_mmhr, Area_km2, last_storm_hr, 
                    dNBR_mean, sand_mean, clay_mean, percent_treated, NDVI, 
                    snow_persi, Ksat_mean, Mean_slope)) %>%
  na.omit # remove NAs, RF doesn't take any NAs


# RF model
rf_sr_et <- randomForest(stage_rise_cm ~., 
                          mtry=3, ntree=1000, # mtry started as 3
                          data=srData_et)


# check model performance for % variance explained
rf_sr_et # note % Var explained isn't super high (~38%), that's fine, we're not trying to produce a predictive model, just explain the data

#view crude graph of variable importance rankings
varImpPlot(rf_sr_et) 

# create more customaizable/interpretable variable importance rankings
varImps_sr_et <- rf_sr_et$importance %>%
  as.data.frame # save increased node purity (metric of variable importance) to data frame

stage_rise_et_RF <- varImps_sr_et %>%
  mutate(., normINP = IncNodePurity/max(IncNodePurity)) %>% # IncNodePurity is the metric for regression RFs (continous data)
  rownames_to_column('Predictor') %>%
  mutate(Predictor = as.factor(Predictor)) %>%
  ggplot(aes(x=reorder(Predictor, normINP), y=normINP)) + theme_bw(base_size = 20) +
      geom_bar(stat='identity', color='black', fill='blue') +
      coord_flip() +
  scale_x_discrete(labels = custom_labels) +
      labs(y='Normalized Variable Importance') +
      theme(axis.title.y = element_blank()) 
      #ggtitle('ET Factors influencing stage rise (cm) across fires')

ggsave(here('figures/stage_rise_et.jpg'), stage_rise_et_RF, dpi=800)

```

# 1 variable importance fig for all 3

```{r}

all_varimport <- varImps_sr %>%
  mutate(., normINP = IncNodePurity/max(IncNodePurity)) %>% # IncNodePurity is the metric for regression RFs (continous data)
  rownames_to_column('Predictor') %>%
  mutate(Predictor = as.factor(Predictor),
         Group = 'All')
  
  
benn_varimport <- varImps_sr_benn %>%
  mutate(., normINP = IncNodePurity/max(IncNodePurity)) %>% # IncNodePurity is the metric for regression RFs (continous data)
  rownames_to_column('Predictor') %>%
  mutate(Predictor = as.factor(Predictor),
         Group = 'CPF')

et_varimport <- varImps_sr_et %>%
  mutate(., normINP = IncNodePurity/max(IncNodePurity)) %>% # IncNodePurity is the metric for regression RFs (continous data)
  rownames_to_column('Predictor') %>%
  mutate(Predictor = as.factor(Predictor),
         Group = 'ETF')

varimport_combine <- bind_rows(all_varimport, benn_varimport, et_varimport)

combine_RF_plot <- varimport_combine %>%
  ggplot(aes(x=reorder(Predictor, normINP), 
             y=normINP, 
             color = Group, 
             shape = Group)) +
  theme_bw(base_size = 20) +
      geom_point(size = 12, alpha = 0.85) +
      coord_flip() +
  scale_x_discrete(labels = custom_labels) +
      labs(y='Normalized Variable Importance',
           x = NULL) +
       theme(legend.title = element_blank(),
        legend.position = "bottom",
        text = element_text(size = 28, family = 'Calibri')) +
    scale_color_manual(values = c("grey62", 'black', 'blue')) +
  scale_shape_manual(values = c(17,16, 18)) 

combine_RF_plot

ggsave(here('figures/stage_rise_combine.jpg'), combine_RF_plot, dpi=800)

```




# Stage response (0/1) ~ all sites

```{r}

# RF MODEL OF OCCURENCE OF STAGE RISE ACROSS BENNETT AND ET (BINARY MODEL) --------------

#subset data only to dependent variable and predictors
biData <- allData %>% # subset data to observed stage rise events
  rename(percent_treated = `%treated`) %>%
  select(., c(response_manual, ID, year, MI60_mmhr, Area_km2, 
              time_since_laststorm_hr, dNBR_mean, sand_mean, 
              clay_mean, percent_treated, NDVI, snow_persi, 
              Ksat_mean, Mean_slope)) %>%
  na.omit

rf_bi_all <- randomForest(as.factor(response_manual) ~., # have to do the dependent variable as factor so it performs classification (categorical data) not regression (continuous data)
                          mtry=3, ntree=1000,
                          data=biData)

# check model performance for % variance explained, note the output changes 
rf_bi_all # note: error estimate = ~13%, so 13 percent of 0,1 outcomes are incorrectly predicted. Confusion matrix in output shows which were predicted (in)correctly

# crude graph of variable importance ranks
varImpPlot(rf_bi_all) 

# create more customizable/interpretable viewing variable importance rankings
varImps_bi <- rf_bi_all$importance %>%
  as.data.frame # save increased node purity (metric of variable importance) to data frame

response_all_RF <- varImps_bi %>%
  mutate(., normINP = MeanDecreaseGini/max(MeanDecreaseGini)) %>% # MeanDecreaseGini is the metric for classification RFs (categorical data)
  rownames_to_column('Predictor') %>%
  mutate(Predictor = as.factor(Predictor)) %>%
  ggplot(., aes(x=reorder(Predictor, normINP), y=normINP)) + theme_bw(base_size = 20) +
      geom_bar(stat='identity', color='black', fill='black') +
      coord_flip() +
  scale_x_discrete(labels = custom_labels) +
      labs(y='Normalized Variable Importance') +
      theme(axis.title.y = element_blank())
      #ggtitle('Factors influencing occurence of stage rise')

ggsave(here('figures/response_all_RF.jpg'), response_all_RF, dpi=800)

```

# Stage response (0/1) ~ Benn

```{r}

# RF MODEL OF OCCURENCE OF STAGE RISE ACROSS BENNETT AND ET (BINARY MODEL) --------------

#subset data only to dependent variable and predictors
biData_benn <- allData %>% # subset data to observed stage rise events
  rename(percent_treated = `%treated`) %>%
  select(., c(response_manual, ID, year, MI60_mmhr, Area_km2, 
              time_since_laststorm_hr, dNBR_mean, sand_mean, 
              clay_mean, percent_treated, NDVI, snow_persi, 
              Ksat_mean, Mean_slope)) %>%
  na.omit

rf_bi_benn <- randomForest(as.factor(response_manual) ~., # have to do the dependent variable as factor so it performs classification (categorical data) not regression (continuous data)
                          mtry=3, ntree=1000,
                          data=biData_benn)

# check model performance for % variance explained, note the output changes 
rf_bi_benn # note: error estimate = ~13%, so 13 percent of 0,1 outcomes are incorrectly predicted. Confusion matrix in output shows which were predicted (in)correctly

# crude graph of variable importance ranks
varImpPlot(rf_bi_benn) 

# create more customizable/interpretable viewing variable importance rankings
varImps_bi_benn <- rf_bi_benn$importance %>%
  as.data.frame # save increased node purity (metric of variable importance) to data frame

response_benn_RF <- varImps_bi_benn %>%
  mutate(., normINP = MeanDecreaseGini/max(MeanDecreaseGini)) %>% # MeanDecreaseGini is the metric for classification RFs (categorical data)
  rownames_to_column('Predictor') %>%
  mutate(Predictor = as.factor(Predictor)) %>%
  ggplot(., aes(x=reorder(Predictor, normINP), y=normINP)) + theme_bw(base_size = 20) +
      geom_bar(stat='identity', color='black', fill='grey62') +
      coord_flip() +
  scale_x_discrete(labels = custom_labels) +
      labs(y='Normalized Variable Importance') +
      theme(axis.title.y = element_blank())
      #ggtitle('CP Factors influencing occurence of stage rise')

ggsave(here('figures/response_benn_RF.jpg'), response_benn_RF, dpi=800)

```

# Stage rise (0/1) ~ ET 

```{r}

# RF MODEL OF OCCURENCE OF STAGE RISE ACROSS BENNETT AND ET (BINARY MODEL) --------------

#subset data only to dependent variable and predictors
biData_et <- allData %>% # subset data to observed stage rise events
  rename(percent_treated = `%treated`) %>%
  select(., c(response_manual, ID, year, MI60_mmhr, Area_km2, 
              time_since_laststorm_hr, dNBR_mean, sand_mean, 
              clay_mean, percent_treated, NDVI, snow_persi, 
              Ksat_mean, Mean_slope)) %>%
  na.omit

rf_bi_et <- randomForest(as.factor(response_manual) ~., # have to do the dependent variable as factor so it performs classification (categorical data) not regression (continuous data)
                          mtry=3, ntree=1000,
                          data=biData_et)

# check model performance for % variance explained, note the output changes 
rf_bi_et # note: error estimate = ~13%, so 13 percent of 0,1 outcomes are incorrectly predicted. Confusion matrix in output shows which were predicted (in)correctly

# crude graph of variable importance ranks
varImpPlot(rf_bi_et) 

# create more customizable/interpretable viewing variable importance rankings
varImps_bi_et <- rf_bi_et$importance %>%
  as.data.frame # save increased node purity (metric of variable importance) to data frame

response_et_RF <- varImps_bi_et %>%
  mutate(., normINP = MeanDecreaseGini/max(MeanDecreaseGini)) %>% # MeanDecreaseGini is the metric for classification RFs (categorical data)
  rownames_to_column('Predictor') %>%
  mutate(Predictor = as.factor(Predictor)) %>%
  ggplot(., aes(x=reorder(Predictor, normINP), y=normINP)) + theme_bw(base_size = 20) +
      geom_bar(stat='identity', color='black', fill='blue') +
      coord_flip() +
      labs(y='Normalized Variable Importance') +
  scale_x_discrete(labels = custom_labels) +
      theme(axis.title.y = element_blank()) 
      #ggtitle('ET Factors influencing occurence of stage rise')

ggsave(here('figures/response_ET_RF.jpg'), response_et_RF, dpi=800)

```
