==> code_archive/7c_et_modeling.Rmd <==
---
title: "East Troublesome Hydrology"
author: "Megan Sears"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: yes
    toc_float: yes
    theme: journal
editor_options:
  chunk_output_type: console
---

```{r, include = F}
knitr::opts_chunk$set(echo = F,
                      message = F,
                      fig.width = 14,
                      fig.height = 6)

library(tidyverse)
library(lubridate)
library(here)
library(ggplot2); theme_set(theme_bw(base_size = 24))
library(corrplot)
library(stats)
library(ggcorrplot)
library(dataRetrieval)
library(lme4)

```

# Stage data

```{r}

response <- read_csv(here('./data/final/stage_response_all.csv')) %>%
  filter(site %in% c('p1','p2',
                     'hum','hm',
                     'mum','mpm',
                     'mm_et','mub',
                     'lum','lpm',
                     'lm')) %>%
  select(event_sum_mm, 
         MI60_mmhr, 
         year, 
         start_time, 
         stage_rise_cm, 
         site, response_manual,
         lag2p_hr,
         datetime_maxstage) %>%
  mutate(Date = as.Date(start_time),
         month = month(Date)) %>%
  filter(!month %in% c(5,10))

```

# Predictors

```{r}

# predictors are in 4 different files (not including rain)
inputs1 <- read_csv(here('./data/final_model_inputs/inputs1.csv')) %>%
  filter(site %in% c('p1','p2',
                     'hum','hm',
                     'mum','mpm',
                     'mm_et','mub',
                     'lum','lpm',
                     'lm')) %>%
  select(-c(Shape_Leng, Shape_Area,
            ID, area_km2))

inputs2 <- read_csv(here('./data/final_model_inputs/inputs2.csv')) %>%
  filter(site %in% c('p1','p2',
                     'hum','hm',
                     'mum','mpm',
                     'mm_et','mub',
                     'lum','lpm',
                     'lm')) %>%
  select(-c(count, Shape_Leng,
            Shape_Area, ELONGATION,
            RC_CIRCLE, ID,
            area_km2,
            relative_area,
            relative_elev)) %>%
  distinct()

inputs3 <- read_csv(here('./data/final_model_inputs/pwd.csv')) %>%
  filter(site %in% c('p1','p2',
                     'hum','hm',
                     'mum','mpm',
                     'mm_et','mub',
                     'lum','lpm',
                     'lm')) %>%
  select(site, lag_pwd, Date, WY) %>% 
  filter(Date > '2022-05-31')
  
inputs4 <- read_csv(here('./data/final_model_inputs/hypso.csv')) %>%
    filter(site %in% c('p1','p2',
                     'hum','hm',
                     'mum','mpm',
                     'mm_et','mub',
                     'lum','lpm',
                     'lm'))

# join predictor data (all except rain metrics)
pred <- left_join(inputs2, inputs1, by = 'site') # combine inputs1 and inputs2
pred <- left_join(inputs4, pred, by = 'site') # now add inputs 4 to that
pred <- left_join(inputs3, pred, by = 'site') # now add pred to inputs3

# need to make NDVI better in the df (so it's 1 column)
ndvi <- pred %>%
  select(site, NDVI_2021, NDVI_2022, NDVI_2023) %>%
  pivot_longer(!site, names_to = 'year', values_to = 'NDVI') %>%
  mutate(year = as.numeric(sub("^NDVI_", "", year))) %>%
  #filter(!year == 2021) %>%
  rename(WY = year) %>%
  group_by(site, WY) %>%
  distinct()

# add ndvi back in
pred <- pred %>%
  select(-c(NDVI_2021, NDVI_2022,
            NDVI_2023)) %>%
  left_join(., ndvi, by = c('site','WY')) 

pred <- pred %>%
  select(-c(geology, geo_general,
            AVG_SAND, AVG_SILT, RC_CIRCLE,
            recharge_total_km2,
            hypso25, hypso75, slope30_frac))


```

# Join together

```{r}

# join the response and pred data
join <- left_join(response, pred, by = c('site', 'Date')) %>%
  # mutate(stage_rise_cm = if_else(stage_rise_cm < 2.0, NA, stage_rise_cm)) %>%
  # mutate(stage_rise_cm = if_else(stage_rise_cm > 120, NA, stage_rise_cm))
  rename(treated = `%treated`) %>%
  mutate(treated = if_else(site == 'hm' & Date < '2022-07-26',
                              0, treated)) %>%
  mutate(treated = if_else(site == 'mpm' & Date < '2022-08-01',
                              0, treated)) %>%
  mutate(lag2p_hr = if_else(site == 'mpm', lag2p_hr/60, lag2p_hr))

  

# plot the selected chem with predictors
plot_response <- function(pred_name) {
  join %>%
    ggplot(aes(x = .data[[pred_name]],
               y = stage_rise_cm)) + 
    geom_point(size=3)
}

pred_name <- names(join) # define the predictor names

map(pred_name, ~ plot_response(.x)) 

# look at corr values
response1 <- join %>%
  select(where(is.numeric)) %>% 
  cor(., use = 'complete.obs')

# last, make a corr plot
corrplot(response1,
         method = "square", 
         type = 'upper',
         #order = "FPC", 
         tl.col = "black", 
         tl.cex = 0.75)

ggplot(join, aes(site, MI60_mmhr)) + geom_boxplot()

```

# Model

```{r}
# make site id numeric and filter for responses
join1 <- join %>%
  mutate(site_id = as.numeric(factor(site, levels = unique(site)))) %>%
  filter(response_manual == 'Y') %>%
  group_by(site) %>%
  mutate(event = row_number()) %>%
  mutate(site_eventID = paste0(site,"_",event)) 

# check distribution of response var
hist(join1$stage_rise_cm) # GAMMA

hist(join1$MI60_mmhr)

## --------------------------------------------------------------------- ##
model_mi <- lmer(stage_rise_cm ~ MI60_mmhr + mean_dnbr + treated + avg_vbw_m + (1|site_id), 
              #family=gaussian(link = "identity"),
              data = join1)

library(lmerTest)

summary(model_mi)
hist(residuals(model_mi))
qqnorm(residuals(model_mi))
qqline(residuals(model_mi))

model_glm <- glm(stage_rise_cm ~ MI60_mmhr + site_id + NDVI + treated + avg_vbw_m, 
              #family=Gamma(link = "inverse"),
              data = join1)


summary(model_glm)
predicted_values <- predict(model_glm)
results <- data.frame(actual = join1$stage_rise_cm, predicted = predicted_values)

ggplot(results, aes(x = actual, y = predicted)) +
  geom_point() +
  geom_abline(intercept = 0, slope = 1, color = "red", linetype = "dashed") 


results2 <- data.frame(treated = join1$treated, predicted = predicted_values)
ggplot(results2, aes(x = treated, y = predicted)) +
  geom_point() +
  labs(x = "Fraction of treated area",
       y = 'Predicted stage rise (cm)')


results3 <- data.frame(MI60_mmhr = join1$MI60_mmhr, predicted = predicted_values)
ggplot(results3, aes(x = MI60_mmhr, y = predicted)) +
  geom_point() +
  labs(x = "MI60 (mm/hr)",
       y = 'Predicted stage rise (cm)')

####################################
# lag to peak
model_glm <- glm(lag2p_hr ~ MI60_mmhr + site_id + NDVI + treated + avg_vbw_m, 
              #family=Gamma(link = "inverse"),
              data = join1)


summary(model_glm)

ggplot(join1, aes(avg_vbw_m, lag2p_hr)) + geom_point() + scale_y_log10()

predicted_values <- predict(model_glm)
results2 <- data.frame(vbw = join1$avg_vbw_m, predicted = predicted_values)
ggplot(results2, aes(x = vbw, y = predicted)) +
  geom_point() +
  labs(x = "Valley bottom width",
       y = 'Predicted lag to peak (hr)')

predicted_values <- predict(model_glm)
results2 <- data.frame(ndvi = join1$NDVI, predicted = predicted_values)
ggplot(results2, aes(x = ndvi, y = predicted)) +
  geom_point() +
  labs(x = "NDVI",
       y = 'Predicted lag to peak (hr)')


```

# rain figure

```{r}

rain <- read_csv('/Users/megansears/Documents/Repos/post-fire_rainfall/data/final/mrms/all_mrms_intensities.csv') %>%
    filter(ID %in% c('p1','p2',
                     'hum','hm',
                     'mum','mpm',
                     'mm_et','mub',
                     'lum','lpm',
                     'lm')) %>%
  mutate(ID = if_else(ID == 'mm_et', 'mm', ID)) %>%
  mutate(mulch = if_else (ID %in% c('hm', 'lm', 'lpm', 'mm', 'mpm'), 'Mulch', 'No treatment')) %>% 
  mutate(ID = factor(ID, levels = c('p1','p2','hum','hm','mum','mpm','mm','mub','lum','lpm', 'lm'))) %>%
  mutate(Date = as.Date(start_time),
         month = month(Date)) %>%
  filter(!month %in% c(5,10)) %>%
  mutate(year = year(Date)) %>%
  filter(year %in% c(2022, 2023))




library(grid) # Required for unit()

ggplot(rain, aes(x = ID, y = MI60_mmhr, fill = mulch)) + 
  geom_boxplot() +
  labs(x = 'Site', y = 'MI60 (mm/hr)') + 
  scale_y_log10() +
  scale_fill_manual(name = '', values = c("Mulch" = "goldenrod3", "No treatment" = "white")) +
  theme(
    legend.position = "bottom",
    legend.direction = "horizontal",
    #legend.key.size = unit(0.5, "cm"), # Adjust the size of the legend keys
    #legend.text = element_text(size = 8), # Adjust the size of the legend text
    plot.margin = unit(c(0.25, 0.25, 0.25, 0.25), "cm"), # Adjust the plot margins (top, right, bottom, left)
    legend.box.margin = margin(-25, 0, 0, 0)) # Adjust the legend box margin


rain_stats <- rain %>%
  group_by(ID) %>% 
  summarize(
    sum = sum(event_sum_mm),
    mean_value = mean(MI60_mmhr),
    sd_value = sd(MI60_mmhr))

response_stats <- response %>%
  filter(response_manual == 'Y') %>% 
  group_by(site) %>% 
  summarize(
    mean_value = mean(stage_rise_cm),
    sd_value = sd(stage_rise_cm),
    max = max(stage_rise_cm))


```

# stage rise fig

```{r}
response1 <- response %>% 
  mutate(mulch = if_else (site %in% c('hm', 'lm', 'lpm', 'mm', 'mpm'), 'Mulch', 'No treatment')) %>% 
  mutate(site = factor(site, levels = c('p1','p2','hum','hm','mum','mpm','mm','mub','lum','lpm', 'lm'))) %>%
  filter(response_manual == 'Y')

# response
ggplot(response1, aes(x = site, y = stage_rise_cm, fill = mulch)) + 
  geom_boxplot() +
  labs(x = 'Site', y = 'Stage rise (cm)') + 
  scale_y_log10() +
  scale_fill_manual(name = '', values = c("Mulch" = "goldenrod3", "No treatment" = "white")) +
  theme(
    legend.position = "bottom",
    legend.direction = "horizontal",
    #legend.key.size = unit(0.5, "cm"), # Adjust the size of the legend keys
    #legend.text = element_text(size = 8), # Adjust the size of the legend text
    plot.margin = unit(c(0.25, 0.25, 0.25, 0.25), "cm"), # Adjust the plot margins (top, right, bottom, left)
    legend.box.margin = margin(-25, 0, 0, 0)) # Adjust the legend box margin

# plot cam and response data together
response2 <- response1 %>%
  select(site, MI60_mmhr, stage_rise_cm)

ggplot(response1, aes(MI60_mmhr, stage_rise_cm, color=mulch)) + geom_point(size=4) +
  scale_y_log10() +
  labs(y='Stage rise (cm)', x = 'MI60 (mm/hr)') +
  scale_color_manual(name = '', values = c("Mulch" = "goldenrod3", "No treatment" = "black")) +
  theme(
    legend.position = "bottom",
    legend.direction = "horizontal",
    #legend.key.size = unit(0.5, "cm"), # Adjust the size of the legend keys
    #legend.text = element_text(size = 8), # Adjust the size of the legend text
    plot.margin = unit(c(0.25, 0.25, 0.25, 0.25), "cm"), # Adjust the plot margins (top, right, bottom, left)
    legend.box.margin = margin(-25, 0, 0, 0)) # Adjust the legend box margin

```

# camera data

```{r}

camera <- read_csv('/Users/megansears/Documents/Repos/post-fire_rainfall/data/et_camera_stream_data.csv') %>%
  mutate(mulch = if_else(Site %in% c('lm', 'lpm', 'lum', 'mm'), 'Mulch', 'No treatment')) 
  

ggplot(camera, aes(Site, stage_rise_cm, fill=mulch)) + geom_boxplot() +
  scale_fill_manual(name = '', values = c("Mulch" = "goldenrod3", "No treatment" = "white")) +
  labs(x='Site', y='Stage rise (cm)') +
    theme(
    legend.position = "bottom",
    legend.direction = "horizontal",
    #legend.key.size = unit(0.5, "cm"), # Adjust the size of the legend keys
    #legend.text = element_text(size = 8), # Adjust the size of the legend text
    plot.margin = unit(c(0.25, 0.25, 0.25, 0.25), "cm"), # Adjust the plot margins (top, right, bottom, left)
    legend.box.margin = margin(-25, 0, 0, 0)) # Adjust the legend box margin


response_stats_cam <- camera %>%
  drop_na(stage_rise_cm) %>% 
  group_by(Site) %>% 
  summarize(
    mean_value = mean(stage_rise_cm),
    sd_value = sd(stage_rise_cm),
    max = max(stage_rise_cm))


camera1 <- camera %>%
  select(ID=Site, Date, Time_first_water, stage_rise_cm) %>%
  mutate(Date = mdy(Date)) %>%
  mutate(datetime = ymd_hm(paste(Date, Time_first_water))) %>%
  drop_na()
  
rain1 <- rain %>%
  select(start_time, end_time, ID, MI60_mmhr) %>%
  filter(ID %in% c('lm', 'lpm', 'lum', 'mm'))

result <- camera1 %>%
  inner_join(rain1, by = "ID") %>%
  filter(datetime >= start_time & datetime <= (end_time + 6 * 60 * 60)) 

response2 <- response1 %>%
  select(site, stage_rise_cm, MI60_mmhr)

result <- result %>%
  select(site=ID, stage_rise_cm, MI60_mmhr)

response3 <- bind_rows(response2, result) %>%
  mutate(mulch = if_else (site %in% c('hm', 'lm', 'lpm', 'mm', 'mpm'), 'Mulch', 'No treatment')) 

ggplot(response3, aes(MI60_mmhr, stage_rise_cm, color=mulch)) + geom_point(size=4) +
  scale_y_log10() +
  labs(y='Stage rise (cm)', x = 'MI60 (mm/hr)') +
  scale_color_manual(name = '', values = c("Mulch" = "goldenrod3", "No treatment" = "black")) +
  theme(
    legend.position = "bottom",
    legend.direction = "horizontal",
    #legend.key.size = unit(0.5, "cm"), # Adjust the size of the legend keys
    #legend.text = element_text(size = 8), # Adjust the size of the legend text
    plot.margin = unit(c(0.25, 0.25, 0.25, 0.25), "cm"), # Adjust the plot margins (top, right, bottom, left)
    legend.box.margin = margin(-25, 0, 0, 0)) # Adjust the legend box margin


```
