==> code_archive/8_figures.Rmd <==
---
title: "Create figures"
author: "Megan Sears"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: yes
    toc_float: yes
    theme: journal
editor_options:
  chunk_output_type: console
---

```{r, include = F}

knitr::opts_chunk$set(echo = F,
                      message = F,
                      fig.width = 12,
                      fig.height = 6)

library(tidyverse)
library(lubridate)
library(plotly)
library(kableExtra)
library(here)
library(ggplot2); theme_set(theme_bw(base_size = 16))
library(cowplot)
```

# Read in data

```{r}

response_all <- read_csv('model_variables.csv') %>%
  mutate(response_manual = ifelse(is.na(response_manual), 
                                  0, 
                                  response_manual))
  

events_all <- read_csv('mrms_intensities.csv') %>%
  filter(!ID == 'mmu')

```

# Precip figures

```{r}

precip_box <- events_all %>%
  #mutate(site = factor(site, levels = c('me', 'mm', 'mw', 'ue', 'um', 'uw',
  #                                      'p1','p2','hum','hm','mum','mpm','mub'))) %>%
  ggplot(., aes(x = ID, MI30_mmhr)) +
  geom_boxplot() +
  facet_wrap(~year)
  
precip_box

precip_maxmi <- events_all %>%
  group_by(ID, year) %>%
  summarize(max_MI30 = max(MI30_mmhr)) %>%
    mutate(ID = factor(ID, levels = c('me', 'mm', 'mw', 'ue', 'um', 'uw',
                                        'p1','p2','hum','hm','mum','mpm','mub'))) %>%
  ggplot(., aes(x=ID, y=max_MI30, 
                color = as.character(year),
                shape = as.character(year))) +
  geom_point()

precip_maxmi

precip_eventsum <- events_all %>%
    mutate(ID = factor(ID, levels = c('me', 'mm', 'mw', 'ue', 'um', 'uw',
                                        'p1','p2','hum','hm','mum','mpm','mub'))) %>%
  ggplot(., aes(x=ID, y=event_sum_mm, 
                color = as.character(year),
                shape = as.character(year))) +
  geom_boxplot()

precip_eventsum

# duration
precip_duration <- events_all %>%
    mutate(ID = factor(ID, levels = c('me', 'mm', 'mw', 'ue', 'um', 'uw',
                                        'p1','p2','hum','hm','mum','mpm','mub'))) %>%
  ggplot(., aes(x=ID, y=duration_hr,
                color = as.character(year),
                shape = as.character(year))) +
  geom_boxplot()

precip_duration

rain_sum <- events_all %>%
  group_by(ID, year) %>%
  summarize(totalP = sum(event_sum_mm))

last_storm <- events_all %>%
  arrange(start_time) %>%
  group_by(ID, year) %>%
  mutate(last_storm_hr =as.numeric(start_time - lag(end_time))) %>%
  select(ID, event, last_storm_hr)

response_all <- left_join(response_all, last_storm, by = c('ID', 'event', 'year'))
write_csv(response_all, 'model_variables_update.csv')

#ggsave(here('figures/et_stage_box.jpg'), et_stage_rise_box, dpi=800)

```

# Predictor figures

Area_km2, dNBR mean, %treated, NDVI, snow persistence, time since last storm?

```{r}
# prep for predictor figs
# categorize dnbr values
map_value_colors <- function(value) {
  case_when(
    between(value, -100, 99) ~ "Unburned",
    between(value, 100, 269) ~ "Low",
    between(value, 270, 439) ~ "Moderate-Low",
    between(value, 440, 659) ~ "Moderate-High",
    between(value, 660, 1300) ~ "High",
    TRUE ~ "Other"  # Default category for values outside specified ranges
  )
}

response_all <- response_all %>%
  mutate(value_colors = map_value_colors(dNBR_mean))

# Pick colors for burn severity colors
category_colors <- c(Unburned = "darkgreen", Low = "#E5E539", `Moderate-Low` = "#E5A039",
                     `Moderate-High` = "#FC7C07", High = "red")

# Order of sites for figure
site_order <- c('me', 'mm', 'mw', 'ue', 'um', 'uw', 'p1', 'p2', 'hum', 'hm', 'mum', 'mpm', 'mub')

# Convert the "site" variable to a factor with the desired order
response_all$site <- factor(response_all$site, levels = site_order)

# area figure showing burn severity
area_fig <- response_all %>%
  ggplot(aes(x = site, y = Area_km2, color = value_colors)) +
  geom_point(size = 14, shape = 18) +
  scale_color_manual(values = category_colors, breaks = names(category_colors),
                     labels = c("Unburned", "Low", "Moderate-Low", "Moderate-High", "High")) +
  labs(color = NULL,
       x = NULL,
       y = expression('Area (' ~ km^2 ~ ')')) +
  theme(
    text = element_text(family = "Times New Roman", size = 30),
    legend.position = "none")

area_fig

ggsave(here('figures/area_fig.jpg'), area_fig, dpi=800)

# NDVI showing burn severity
ndvi_fig <- response_all %>%
  ggplot(aes(x = site, y =NDVI, color = value_colors, shape = as.character(year))) +
  geom_point(size = 10) +
  scale_color_manual(values = category_colors, breaks = names(category_colors),
                     labels = c("Unburned", "Low", "Moderate-Low", "Moderate-High", "High")) +
  labs(shape = NULL,
    color = NULL,
       x = NULL,
       y = expression('NDVI')) +
  theme(
    text = element_text(family = "Times New Roman", size = 34),
    legend.position = "none")

ndvi_fig

ggsave(here('figures/ndvi_fig.jpg'), ndvi_fig, dpi=800)

# NDVI averages by year

ndvi_yr <- response_all %>%
  filter(!site %in% c('mub', 'p1')) %>%
  group_by(fire, year) %>%
  summarize(mean_NDVI = mean(NDVI)) %>%
  ggplot(., aes(year, mean_NDVI, color = as.character(fire),
                shape = as.character(fire))) + 
  geom_point(size = 14) +
  scale_x_continuous(breaks=c(2021, 2022, 2023)) +
    scale_shape_manual(labels = c('CPF', 'ETF'), 
                       values=c(19,18)) +
  labs (x = 'Year', y = 'Mean NDVI') +
  scale_color_manual(values = c('black', 'blue'),
                                labels = c('CPF', 'ETF')) +
    theme(
    text = element_text(family = "Times New Roman", size = 34),
    legend.position = "bottom",
    legend.title=element_blank())


ndvi_yr

ggsave(here('figures/ndvi_yr_fig.jpg'), ndvi_yr, dpi=800)

# %treated fig showing burn severity
treat_fig <- response_all %>%
  ggplot(aes(x = site, y = `%treated`, color = value_colors)) +
  geom_point(size = 14, shape = 18) +
  scale_color_manual(values = category_colors, breaks = names(category_colors),
                     labels = c("Unburned", "Low", "Moderate-Low", "Moderate-High", "High")) +
  labs(color = NULL,
       x = NULL,
       y = '% area with mulch treatment') +
  theme(
    text = element_text(family = "Times New Roman", size = 30),
    legend.position = "none")

treat_fig

ggsave(here('figures/treatment_fig.jpg'), treat_fig, dpi=800)

# SP fig showing burn severity
sp_fig <- response_all %>%
  ggplot(aes(x = site, y = snow_persi, color = value_colors)) +
  geom_point(size = 14, shape = 18) +
  scale_color_manual(values = category_colors, breaks = names(category_colors),
                     labels = c("Unburned", "Low", "Moderate-Low", "Moderate-High", "High")) +
  labs(color = NULL,
       x = NULL,
       y = 'Snow persistence (%)') +
  theme(
    text = element_text(family = "Times New Roman", size = 30),
    legend.position = "none")

sp_fig

ggsave(here('figures/snowper_fig.jpg'), sp_fig, dpi=800)

# last storm
laststorm_fig <- response_all %>%
  ggplot(aes(x = site, y = last_storm_hr, fill = value_colors)) +
  geom_boxplot() +
  scale_fill_manual(values = category_colors, breaks = names(category_colors),
                     labels = c("Unburned", "Low", "Moderate-Low", "Moderate-High", "High")) +
  labs(color = NULL,
       x = NULL,
       y = 'Hours since last storm') +
  theme(
    text = element_text(family = "Times New Roman", size = 34),
    legend.position = "none")

laststorm_fig

ggsave(here('figures/laststorm_fig.jpg'), laststorm_fig, dpi=800)

```

# Stage rise figure between two sites

```{r}

# Stage rise vs MI for CPF vs ETF
scatter_mi60_vs <- response_all %>%
  mutate(fire_name = if_else(fire == 1,
                             "CPF", "ETF"),
         burn_cat = if_else(fire_name == 'ETF' & value_colors == 'Unburned',
                            'ETF unburned',
                            fire_name),
         burn_yn = if_else(value_colors == 'Unburned',
                           'Unburned',
                           'Burned')) %>%
  ggplot(., aes(MI60_mmhr, stage_rise_cm, 
                         color=burn_cat,
                shape = burn_cat)) +
  geom_point(size=12, alpha = 0.75) +
  scale_color_manual(values=c("black", "blue", 'orchid4')) +
  scale_shape_manual(values=c(19,18,15)) +
  xlab('MI60 (mm/hr)') +
  ylab("Stage rise (cm)") +
  theme(legend.title=element_blank()) +
  theme(legend.position="bottom",
        text = element_text(size=30,
                            family = 'Times New Roman'))


scatter_mi60_vs

ggsave(here('figures/scatter_stage_mi.jpg'), scatter_mi60_vs, dpi=800)

```

# Stage rise vs MI by year

```{r}
# Benn 2021-2023 MI vs stage rise for years
scatter_mi_year_benn <- response_all %>%
  mutate(fire_name = if_else(fire == 1, "CPF", "ETF")) %>%
  filter(fire_name == 'CPF') %>% # response_manual == 1 ADD THIS BACK IN
  ggplot(., aes(MI60_mmhr, stage_rise_cm, color = as.character(year), shape = as.character(year))) +
  geom_point(size = 14, alpha = 0.75) +
  scale_color_manual(values = c("#009E73", "gray62", '#0072B2')) +
  xlab('MI60 (mm/hr)') +
  ylab("Stage rise (cm)") +
  theme(legend.title = element_blank()) +
  theme(legend.position = "bottom",
        text = element_text(size = 34, family = 'Times New Roman'))

scatter_mi_year_benn

ggsave(here('figures/scatter_stage_year_benn.jpg'), scatter_mi_year_benn, dpi=800)

# ET 2022 vs 2023 MI vs stage rise for years - TAKE OUT BURN SITES
scatter_mi_year_et <- response_all %>%
  mutate(fire_name = if_else(fire == 1, "CPF", "ETF"),
         burn_cat = if_else(fire_name == 'ETF' & value_colors == 'Unburned',
                            'ETF unburned',
                            fire_name),
         burn_yr = if_else(burn_cat == 'ETF unburned',
                           'Unburned',
                           as.character(year))) %>%
  filter( 
         fire_name == 'ETF') %>%
  ggplot(., aes(MI60_mmhr, stage_rise_cm, 
                color = burn_yr, 
                shape = burn_yr)) +
  geom_point(size = 14, alpha = 0.75) +
  scale_color_manual(values = c("gray62", '#0072B2', 'orchid4')) +
  scale_shape_manual(values = c(17,15, 18)) +
  xlab('MI60 (mm/hr)') +
  ylab("Stage rise (cm)") +
  ylim(0,130) +
  theme(legend.title = element_blank(),
        legend.position = "bottom",
        text = element_text(size = 34, family = 'Times New Roman'))

scatter_mi_year_et

ggsave(here('figures/scatter_stage_year_et.jpg'), scatter_mi_year_et, dpi=800)

```

# Random stats

```{r}

sum <- response_all %>%
  group_by(fire, year) %>%
  summarize(sum_event = sum(response_manual),
            count = n()) %>%
  mutate(percent_response = (sum_event/count)*100)

benn_2023 <- response_all %>%
  filter(fire == 1,
         year == 2023)

```
